refactor: create Models, types into OneSection
