+++ client/src/Models/FlatModel.ts
@@ -1,0 +1,11 @@
+export interface IFlat {
+  id: number,
+  floor: number,
+  pos_on_floor: number,
+  price: number,
+  rooms: number,
+  area_total: number,
+  area_kitchen: number,
+  area_live: number,
+  layout_image: string,
+}
