+++ client/src/components/OneSection/types.ts
@@ -1,18 +1,3 @@
-export interface IFlat {
-  id: number,
-  floor: number,
-  pos_on_floor: number,
-  price: number,
-  rooms: number,
-  area_total: number,
-  area_kitchen: number,
-  area_live: number,
-  layout_image: string,
-}
-
-export interface ISection {
-  flats: IFlat[],
-}
 
 export interface IFilterParams {
   min_floor?: number,
