+++ client/src/store/API/section.api.ts
@@ -1,7 +1,7 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
 import { SERVER_URL } from '../../const/const';
 
-import { IFlat } from '../../types/types';
+import { IFlat } from '../../Models/FlatModel';
 
 
 export const sectionApi = createApi({
